+++ history.md
@@ -2,6 +2,7 @@
 
 - Add `events.race`.
 - Change default `suggest.triggerCompletionWait` to 50.
+- Support trigger completion after indent fix.
 
 # 2022-02-14